requirements/requirements.gateway.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
transformers
protobuf
sentencepiece
tiktoken
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp